requirements.txt
gradio>=3.0.0
huggingface_hub>=0.13.3
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.3.1
matplotlib>=3.4.2
loguru>=0.5.3
structlog>=21.5.0
llama-cpp-python>=0.1.59
jupyter
docker
